Annotation data for the ‘IlluminaHumanMethylationEPIC’ micro array.

This package is based on the file https://support.illumina.com/content/dam/illumina-support/documents/downloads/productfiles/methylationepic/MethylationEPIC%20v2%20Files.zip from https://support.illumina.com/array/array_kits/infinium-methylationepic-beadchip-kit/downloads.html.

Additional SNP annotation is generated by the authors (described in SNPs.CommonSingle).

A script for creating the data object in this package is contained in scripts/createAnno.R.

Format

An object of class IlluminaMethylationAnnotation for IlluminaHumanMethylationEPIv2Canno.20a1.hg38. All others are of class DataFrame

Details

The following changes/ modifications / addition has been made to the source material.

For the annotation related to UCSC islands, a value of "" has been changed to "OpenSea".

The creation of this object based on the Illumina annotation and additional SNP information (see SNPs.CommonSingle) is contained in the createAnno.R script in the scripts directory.

Contents of the columns are the responsibility of Illumina; see their documentation for details.

Source

See description.

Examples

IlluminaHumanMethylationEPICv2anno.20a1.hg38
#> IlluminaMethylationAnnotation object
#> Annotation
#>   array: IlluminaHumanMethylationEPICv2
#>   annotation: 20a1
#>   genomeBuild: hg38
#> Available annotation
#>   Islands.UCSC
#>   Locations
#>   Manifest
#>   Other
#>   SNPs.141CommonSingle
#>   SNPs.142CommonSingle
#>   SNPs.144CommonSingle
#>   SNPs.146CommonSingle
#>   SNPs.147CommonSingle
#>   SNPs.150CommonSingle
#>   SNPs.151CommonSingle
#>   SNPs.Illumina
#> Defaults
#>   Locations
#>   Manifest
#>   SNPs.151CommonSingle
#>   Islands.UCSC
#>   Other
Locations
#> DataFrame with 930075 rows and 3 columns
#>                              chr       pos      strand
#>                      <character> <integer> <character>
#> cg25324105_BC11            chr19  37692358           +
#> cg25383568_TC11            chr19  38727081           -
#> cg25455143_BC11            chr19   1591515           -
#> cg25459778_BC11            chr16   1614581           +
#> cg25487775_BC11             chr2 161237458           +
#> ...                          ...       ...         ...
#> ch.22.16108507R_BC21       chr22  17247617           +
#> ch.22.40657F_BC21          chr22  17685590           -
#> ch.22.78028F_BC21          chr22  19374531           -
#> ch.22.17803970R_BC21       chr22  19436447           +
#> ch.22.107403R_TC21         chr22  20144102           +
Manifest
#> DataFrame with 930075 rows and 8 columns
#>                                      Name    AddressA    AddressB
#>                               <character> <character> <character>
#> cg25324105_BC11           cg25324105_BC11     1754126    99753217
#> cg25383568_TC11           cg25383568_TC11    79792482    69667133
#> cg25455143_BC11           cg25455143_BC11    80699190     7659147
#> cg25459778_BC11           cg25459778_BC11    60797262    65710482
#> cg25487775_BC11           cg25487775_BC11     5799427    89606481
#> ...                                   ...         ...         ...
#> ch.22.16108507R_BC21 ch.22.16108507R_BC21    88698463            
#> ch.22.40657F_BC21       ch.22.40657F_BC21    51795249            
#> ch.22.78028F_BC21       ch.22.78028F_BC21    43681841            
#> ch.22.17803970R_BC21 ch.22.17803970R_BC21    17741188            
#> ch.22.107403R_TC21     ch.22.107403R_TC21    89738435            
#>                                   ProbeSeqA              ProbeSeqB        Type
#>                                 <character>            <character> <character>
#> cg25324105_BC11      ATTTATAAACTAATAACCCA.. GTTTATAAACTAATAACCCG..           I
#> cg25383568_TC11      AAACCAAAAAAATAACAAAC.. AAACCGAAAAAATAACAAAC..           I
#> cg25455143_BC11      ATAAAAAAAAATATACAACT.. ATAAAAAAAAATATACGACT..           I
#> cg25459778_BC11      AAAAATTTAAAACAAACAAC.. AAAAATTTAAAACAAACAAC..           I
#> cg25487775_BC11      AAAAACAACCTAAAAAACAA.. AAAAACAACCTAAAAAACAA..           I
#> ...                                     ...                    ...         ...
#> ch.22.16108507R_BC21 CTATTCTTTTTACTTAAACA..                                 II
#> ch.22.40657F_BC21    AATAAATAAACAAAATATAA..                                 II
#> ch.22.78028F_BC21    ACAATTCCTAAAAACTAACA..                                 II
#> ch.22.17803970R_BC21 CAACCAAAATTAAAATTAAT..                                 II
#> ch.22.107403R_TC21   CTTTCTCTTTAAAACCAATA..                                 II
#>                         NextBase       Color
#>                      <character> <character>
#> cg25324105_BC11                A         Red
#> cg25383568_TC11                C         Grn
#> cg25455143_BC11                T         Red
#> cg25459778_BC11                C         Grn
#> cg25487775_BC11                A         Red
#> ...                          ...         ...
#> ch.22.16108507R_BC21                        
#> ch.22.40657F_BC21                           
#> ch.22.78028F_BC21                           
#> ch.22.17803970R_BC21                        
#> ch.22.107403R_TC21                          
data(SNPs.Illumina)  # data object too large, need to be loaded manually by data()
Islands.UCSC
#> DataFrame with 930075 rows and 2 columns
#>                                Islands_Name Relation_to_Island
#>                                 <character>        <character>
#> cg25324105_BC11      chr19:37691892-37692..             Island
#> cg25383568_TC11      chr19:38726890-38727..             Island
#> cg25455143_BC11      chr19:1591428-159162..             Island
#> cg25459778_BC11       chr16:1610053-1615094             Island
#> cg25487775_BC11      chr2:161237949-16123..              Shore
#> ...                                     ...                ...
#> ch.22.16108507R_BC21                                   OpenSea
#> ch.22.40657F_BC21                                      OpenSea
#> ch.22.78028F_BC21                                      OpenSea
#> ch.22.17803970R_BC21 chr22:19431718-19432..              Shelf
#> ch.22.107403R_TC21   chr22:20142873-20143..              Shore
data(Other)  # data object too large, need to be loaded manually by data()