SNP annotation from various versions of dbSNP as represented on UCSC Common SNP table.

SNP annotation from various versions of dbSNP as represented on UCSC Common SNP table. Overlap is based on genomic mappings from the annotation package.

Format

An object of class DataFrame. Rownames are CpG identifiers. There are 6 columns Probe_rs, Probe_maf, CpG_rs, CpG_maf, SBE_rs, and SBE_maf. ‘Probe’ indicates a SNP in the probe, ‘CpG’ a SNP in the CpG site and ‘SBE’ in the single base extension site. The _rs gives the SNP RS identifier and the _maf gives the minor allele frequency.

Details

In addition to the SNP information provided by Illumina, we have added independent information on the overlap of the EPIC (v2.0) array with various versions of dbSNP. The overlap is based on the mappings of the array to the hg38 genome provided by Illumina. As dbSNP we have used the ‘Common’ table from UCSC (ie. ‘snp151Common’). This track contains variants from dbSNP which have a minor allele frequency (MAF) of greater than 1 percent (specifically, this requires dbSNP to actually contain MAF information). Furthermore, we only kept variants marked as ‘single’ (ie. standard single nucleotide changes, but not insertions or deletions). Scripts for retrieving the UCSC dbSNP table and doing the overlap are contained in the scripts directory. The variants are described in 6 different columns. Probe_rs tells us the RS number (SNP ID number) for a SNP overlapping the probe, and Probe_maf is the minor allelle frequency for the SNP (in case multiple SNPs overlap, only one is recorded). Similarly, CpG_rs describe SNPs overlapping the CpG site and SBE_rs the single base extension of the measured methylation loci.

Source

UCSC Common SNP Table.

Examples

SNPs.141CommonSingle
#> DataFrame with 930075 rows and 6 columns
#>                         Probe_rs Probe_maf      CpG_rs   CpG_maf      SBE_rs
#>                      <character> <numeric> <character> <numeric> <character>
#> cg25324105_BC11               NA        NA          NA        NA          NA
#> cg25383568_TC11               NA        NA          NA        NA          NA
#> cg25455143_BC11               NA        NA          NA        NA          NA
#> cg25459778_BC11               NA        NA          NA        NA          NA
#> cg25487775_BC11               NA        NA          NA        NA          NA
#> ...                          ...       ...         ...       ...         ...
#> ch.22.16108507R_BC21   rs5748987  0.257091          NA        NA          NA
#> ch.22.40657F_BC21             NA        NA          NA        NA          NA
#> ch.22.78028F_BC21             NA        NA          NA        NA          NA
#> ch.22.17803970R_BC21          NA        NA          NA        NA          NA
#> ch.22.107403R_TC21            NA        NA          NA        NA          NA
#>                        SBE_maf
#>                      <numeric>
#> cg25324105_BC11             NA
#> cg25383568_TC11             NA
#> cg25455143_BC11             NA
#> cg25459778_BC11             NA
#> cg25487775_BC11             NA
#> ...                        ...
#> ch.22.16108507R_BC21        NA
#> ch.22.40657F_BC21           NA
#> ch.22.78028F_BC21           NA
#> ch.22.17803970R_BC21        NA
#> ch.22.107403R_TC21          NA
SNPs.142CommonSingle
#> DataFrame with 930075 rows and 6 columns
#>                         Probe_rs Probe_maf      CpG_rs   CpG_maf      SBE_rs
#>                      <character> <numeric> <character> <numeric> <character>
#> cg25324105_BC11               NA        NA          NA        NA          NA
#> cg25383568_TC11               NA        NA          NA        NA          NA
#> cg25455143_BC11               NA        NA          NA        NA          NA
#> cg25459778_BC11               NA        NA          NA        NA          NA
#> cg25487775_BC11               NA        NA          NA        NA          NA
#> ...                          ...       ...         ...       ...         ...
#> ch.22.16108507R_BC21   rs5748987  0.250799          NA        NA          NA
#> ch.22.40657F_BC21    rs143942158  0.015575          NA        NA          NA
#> ch.22.78028F_BC21             NA        NA          NA        NA          NA
#> ch.22.17803970R_BC21          NA        NA          NA        NA          NA
#> ch.22.107403R_TC21            NA        NA          NA        NA          NA
#>                        SBE_maf
#>                      <numeric>
#> cg25324105_BC11             NA
#> cg25383568_TC11             NA
#> cg25455143_BC11             NA
#> cg25459778_BC11             NA
#> cg25487775_BC11             NA
#> ...                        ...
#> ch.22.16108507R_BC21        NA
#> ch.22.40657F_BC21           NA
#> ch.22.78028F_BC21           NA
#> ch.22.17803970R_BC21        NA
#> ch.22.107403R_TC21          NA
SNPs.144CommonSingle
#> DataFrame with 930075 rows and 6 columns
#>                         Probe_rs Probe_maf      CpG_rs   CpG_maf      SBE_rs
#>                      <character> <numeric> <character> <numeric> <character>
#> cg25324105_BC11               NA        NA          NA        NA          NA
#> cg25383568_TC11               NA        NA          NA        NA          NA
#> cg25455143_BC11               NA        NA          NA        NA          NA
#> cg25459778_BC11               NA        NA          NA        NA          NA
#> cg25487775_BC11               NA        NA          NA        NA          NA
#> ...                          ...       ...         ...       ...         ...
#> ch.22.16108507R_BC21   rs5748987  0.250799          NA        NA          NA
#> ch.22.40657F_BC21    rs143942158  0.015575          NA        NA          NA
#> ch.22.78028F_BC21             NA        NA          NA        NA          NA
#> ch.22.17803970R_BC21          NA        NA          NA        NA          NA
#> ch.22.107403R_TC21            NA        NA          NA        NA          NA
#>                        SBE_maf
#>                      <numeric>
#> cg25324105_BC11             NA
#> cg25383568_TC11             NA
#> cg25455143_BC11             NA
#> cg25459778_BC11             NA
#> cg25487775_BC11             NA
#> ...                        ...
#> ch.22.16108507R_BC21        NA
#> ch.22.40657F_BC21           NA
#> ch.22.78028F_BC21           NA
#> ch.22.17803970R_BC21        NA
#> ch.22.107403R_TC21          NA
SNPs.146CommonSingle
#> DataFrame with 930075 rows and 6 columns
#>                         Probe_rs Probe_maf      CpG_rs   CpG_maf      SBE_rs
#>                      <character> <numeric> <character> <numeric> <character>
#> cg25324105_BC11               NA        NA          NA        NA          NA
#> cg25383568_TC11               NA        NA          NA        NA          NA
#> cg25455143_BC11               NA        NA          NA        NA          NA
#> cg25459778_BC11               NA        NA          NA        NA          NA
#> cg25487775_BC11               NA        NA          NA        NA          NA
#> ...                          ...       ...         ...       ...         ...
#> ch.22.16108507R_BC21   rs5748987  0.250799          NA        NA          NA
#> ch.22.40657F_BC21    rs143942158  0.015575          NA        NA          NA
#> ch.22.78028F_BC21             NA        NA          NA        NA          NA
#> ch.22.17803970R_BC21          NA        NA          NA        NA          NA
#> ch.22.107403R_TC21            NA        NA          NA        NA          NA
#>                        SBE_maf
#>                      <numeric>
#> cg25324105_BC11             NA
#> cg25383568_TC11             NA
#> cg25455143_BC11             NA
#> cg25459778_BC11             NA
#> cg25487775_BC11             NA
#> ...                        ...
#> ch.22.16108507R_BC21        NA
#> ch.22.40657F_BC21           NA
#> ch.22.78028F_BC21           NA
#> ch.22.17803970R_BC21        NA
#> ch.22.107403R_TC21          NA
SNPs.147CommonSingle
#> DataFrame with 930075 rows and 6 columns
#>                         Probe_rs Probe_maf      CpG_rs   CpG_maf      SBE_rs
#>                      <character> <numeric> <character> <numeric> <character>
#> cg25324105_BC11               NA        NA          NA        NA          NA
#> cg25383568_TC11               NA        NA          NA        NA          NA
#> cg25455143_BC11               NA        NA          NA        NA          NA
#> cg25459778_BC11               NA        NA          NA        NA          NA
#> cg25487775_BC11               NA        NA          NA        NA          NA
#> ...                          ...       ...         ...       ...         ...
#> ch.22.16108507R_BC21   rs5748987  0.250799          NA        NA          NA
#> ch.22.40657F_BC21    rs143942158  0.015575          NA        NA          NA
#> ch.22.78028F_BC21             NA        NA          NA        NA          NA
#> ch.22.17803970R_BC21          NA        NA          NA        NA          NA
#> ch.22.107403R_TC21            NA        NA          NA        NA          NA
#>                        SBE_maf
#>                      <numeric>
#> cg25324105_BC11             NA
#> cg25383568_TC11             NA
#> cg25455143_BC11             NA
#> cg25459778_BC11             NA
#> cg25487775_BC11             NA
#> ...                        ...
#> ch.22.16108507R_BC21        NA
#> ch.22.40657F_BC21           NA
#> ch.22.78028F_BC21           NA
#> ch.22.17803970R_BC21        NA
#> ch.22.107403R_TC21          NA
SNPs.150CommonSingle
#> DataFrame with 930075 rows and 6 columns
#>                         Probe_rs Probe_maf      CpG_rs   CpG_maf      SBE_rs
#>                      <character> <numeric> <character> <numeric> <character>
#> cg25324105_BC11               NA        NA          NA        NA          NA
#> cg25383568_TC11               NA        NA          NA        NA          NA
#> cg25455143_BC11               NA        NA          NA        NA          NA
#> cg25459778_BC11               NA        NA          NA        NA          NA
#> cg25487775_BC11               NA        NA          NA        NA          NA
#> ...                          ...       ...         ...       ...         ...
#> ch.22.16108507R_BC21   rs5748987  0.250799          NA        NA          NA
#> ch.22.40657F_BC21    rs143942158  0.015575          NA        NA          NA
#> ch.22.78028F_BC21             NA        NA          NA        NA          NA
#> ch.22.17803970R_BC21          NA        NA          NA        NA          NA
#> ch.22.107403R_TC21            NA        NA          NA        NA          NA
#>                        SBE_maf
#>                      <numeric>
#> cg25324105_BC11             NA
#> cg25383568_TC11             NA
#> cg25455143_BC11             NA
#> cg25459778_BC11             NA
#> cg25487775_BC11             NA
#> ...                        ...
#> ch.22.16108507R_BC21        NA
#> ch.22.40657F_BC21           NA
#> ch.22.78028F_BC21           NA
#> ch.22.17803970R_BC21        NA
#> ch.22.107403R_TC21          NA
SNPs.151CommonSingle
#> DataFrame with 930075 rows and 6 columns
#>                         Probe_rs Probe_maf      CpG_rs   CpG_maf      SBE_rs
#>                      <character> <numeric> <character> <numeric> <character>
#> cg25324105_BC11               NA        NA          NA        NA          NA
#> cg25383568_TC11               NA        NA          NA        NA          NA
#> cg25455143_BC11               NA        NA          NA        NA          NA
#> cg25459778_BC11               NA        NA          NA        NA          NA
#> cg25487775_BC11               NA        NA          NA        NA          NA
#> ...                          ...       ...         ...       ...         ...
#> ch.22.16108507R_BC21   rs5748987  0.250799          NA        NA          NA
#> ch.22.40657F_BC21    rs143942158  0.015575          NA        NA          NA
#> ch.22.78028F_BC21             NA        NA          NA        NA          NA
#> ch.22.17803970R_BC21          NA        NA          NA        NA          NA
#> ch.22.107403R_TC21            NA        NA          NA        NA          NA
#>                        SBE_maf
#>                      <numeric>
#> cg25324105_BC11             NA
#> cg25383568_TC11             NA
#> cg25455143_BC11             NA
#> cg25459778_BC11             NA
#> cg25487775_BC11             NA
#> ...                        ...
#> ch.22.16108507R_BC21        NA
#> ch.22.40657F_BC21           NA
#> ch.22.78028F_BC21           NA
#> ch.22.17803970R_BC21        NA
#> ch.22.107403R_TC21          NA